requirements.txt
pandas
boto
s3fs
python-dotenv
diagram